Go back we messed up

+++ FTPClient.java
@@ -43,11 +43,7 @@ class FTPClient {
         
 	    while(isOpen && clientgo) {
             sentence = inFromUser.readLine();
-			File newfile = new File(sentence.replace("retr: ", ""));
-			if(newfile.exists()){
-			    System.out.println("File already exists.");
-				break;
-			}
+			
             port = port + 2;
 			ServerSocket welcomeData = new ServerSocket(port);
             outToServer.writeBytes (port + " " + sentence + " " + '\n');
@@ -79,19 +75,21 @@ class FTPClient {
                 if (serverStatus.startsWith("550")) {
                     System.out.println("File not found on server.");
                 } else if (serverStatus.startsWith("200")) {
-                    try {
+                     try {
+                        DataInputStream inData = new DataInputStream(new BufferedInputStream(dataSocket.getInputStream()));
+
                         File newfile = new File(sentence.replace("retr: ", ""));
-						DataInputStream inData = new DataInputStream(new BufferedInputStream(dataSocket.getInputStream()));
-						int letter = -1;
-                      
-                        System.out.println("Downloading file...");
-                        FileOutputStream fos = new FileOutputStream(newfile);
-                            
-                        while ((letter = inData.read()) != -1) {
+                        if (!newfile.exists()) {
+                            System.out.println("Downloading file...");
+                            FileOutputStream fos = new FileOutputStream(newfile);
+                            int letter = -1;
+                            while ((letter = inData.read()) != -1) {
                                 fos.write((char) letter);
+                            }
+                            System.out.println("File successfully downloaded!");
+                        } else {
+                            System.out.println("File already exists.");
                         }
-                        System.out.println("File successfully downloaded!");
-                          
                     } catch(IOException e){
                         System.out.println(e);
                     }
